создал модальное окно без внутренней части(нерабочее окно)
