+++ spotify-react-app/src/pages/RegisterPage/components/EmailModal/EmailConfirmationModal.tsx
@@ -1,0 +1,11 @@
+export const EmailConfirmationModal = (props: any) => {
+    return (
+        <div className="modal">
+            <div className="modal-wrapper">
+                <div className="modal-content">
+                    <button type="button" className="modal-close-button"></button>
+                </div>
+            </div>
+        </div>
+    )
+}
